+++ packages/dnr-rulesets/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Added
+- New `watch` command to watch for changes in the filters
+  and automatically rebuild DNR.
+- New option `--latest-filters` for `load` command to load fresh text filters
+  instead of DNR rulesets and save them to the specified directory.
+
 ## [3.1.0] - 2025-06-06
 
 ### Added